requirements.txt
fastapi==0.128.8
firebase_admin==7.2.0
python-dotenv==1.2.1
pytest==8.4.2
requests==2.32.5
uvicorn==0.39.0
